공지사항 등록, 공지사항 리스트 공지사항 업데이트 수정중 공지사항 리스트 상단 고정 해야함

+++ haruProject/src/main/java/com/example/haruProject/dto/Product.java
@@ -19,7 +19,7 @@ import lombok.ToString;
 public class Product {
 	private int pno;			//상품번호
 	private String pname;		//상품명
-	private double pprice;		//상품가격
+	private int pprice;			//상품가격
 	private String pbuy_store;	//구매
 	private String pimg_main;	//메인이미지
 	private String pbrand;		//브랜드
